rename log_server -> logger
